+++ interfaces.go
@@ -77,6 +77,9 @@ type ReserveData interface {
 
 	GetTradeHistory(fromTime, toTime uint64) (common.AllTradeHistory, error)
 
+	UpdateFetcherConfiguration(common.FetcherConfiguration) error
+	GetAllFetcherConfiguration() (common.FetcherConfiguration, error)
+
 	Run() error
 	RunStorageController() error
 	Stop() error