+++ Changelog.md
@@ -7,6 +7,7 @@
 - add /get-step-function-data API for fetching tokens step function configuration
 - add "creation_time" field to Token info
 - add /btc-feed API (#453)
+- add fetcher configuration APIs (#477)
 
 ### Bug fixes: